+++ main.ts
@@ -13,6 +13,7 @@
 import { red } from "@std/fmt/colors";
 import { initCommand } from "./init.ts";
 import { newLogoCommand } from "./new-logo.ts";
+import { newPwaImagesCommand } from "./new-pwa-images.ts";
 import denoConfig from "./deno.json" with { type: "json" };
 
 const args = Deno.args;
@@ -29,14 +30,16 @@ Crunchwrap App CLI (v${denoConfig.version})
 Usage:
   cw init
   cw new-logo
+  cw new-pwa-images
 
 Options:
   -v, --version  Show version
   -h, --help     Show help
 
 Commands:
-  init      Initialize a new Crunchwrap app
-  new-logo  Generate a new logo for an existing app
+  init            Initialize a new Crunchwrap app
+  new-logo        Generate a new logo with AI for an existing app
+  new-pwa-images  Generate PWA images from an existing image
 `);
   Deno.exit(0);
 }
@@ -50,6 +53,11 @@ if (cmd === "init") {
 
 if (cmd === "new-logo") {
   await newLogoCommand();
+  Deno.exit(0);
+}
+
+if (cmd === "new-pwa-images") {
+  await newPwaImagesCommand();
   Deno.exit(0);
 }